--- HOJIN/WEEK1/.idea/modules.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/.idea/WEEK1.iml" filepath="$PROJECT_DIR$/.idea/WEEK1.iml" />
-    </modules>
-  </component>
-</project>
