--- SEOJIN/WEEK1/SJ_1929.java
@@ -1,37 +1,0 @@
-package Algorithm.DataStructure;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.StringTokenizer;
-
-public class SJ_1929 {
-
-    public static void main(String[] args) throws IOException {
-
-        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        StringTokenizer st = new StringTokenizer(br.readLine());
-
-        int N = Integer.parseInt(st.nextToken());
-        int M = Integer.parseInt(st.nextToken());
-
-        for (int i = N; i <= M; i++) {
-            if (i < 2){
-                continue;
-            }
-            boolean isPrime = true;
-
-            for (int j = 2; j <= Math.sqrt(i) ; j++) {
-                if (i % j == 0){
-                    isPrime = false;
-                    break;
-                }
-            }
-            if(isPrime){
-                System.out.println(i);
-            }
-        }
-
-    }
-
-}
